analytics functionality

+++ services/project/internal/logic/assignment.go
@@ -8,12 +8,12 @@ import (
 	assignmentdata "pms.project/internal/data/assignment"
 )
 
-func (l *Logic) AssginTask(ctx context.Context, userID, taskID string) error {
-	log := l.log.Named("AssginTask").With(
+func (l *Logic) AssignTask(ctx context.Context, userID, taskID string) error {
+	log := l.log.Named("AssignTask").With(
 		zap.String("user_id", userID),
 		zap.String("task_id", taskID),
 	)
-	log.Debug("AssginTask called")
+	log.Debug("AssignTask called")
 
 	task, err := l.Repo.Task.GetByID(ctx, taskID)
 	if err != nil {
@@ -54,10 +54,10 @@ func (l *Logic) UnassignTask(ctx context.Context, userID, taskID string) error {
 	)
 	log.Debug("UnassignTask called")
 
-	// existing, _ := l.Repo.TaskAssignment.Get(ctx, userID, taskID)
-	// if existing == nil {
-	// 	return nil
-	// }
+	existing, _ := l.Repo.TaskAssignment.Get(ctx, userID, taskID)
+	if existing == nil {
+		return nil
+	}
 
 	if err := l.Repo.TaskAssignment.Delete(ctx, assignmentdata.AssignmentData{
 		UserID: userID,
